testando funcionalidades
