+++ modulo3/astromatch/package.json
@@ -10,7 +10,9 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "react-toastify": "^9.0.5",
     "styled-components": "^5.3.5",
+    "sweetalert2": "^11.4.20",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
