+++ modulo3/astromatch/src/components/Home/Home.js
@@ -5,6 +5,8 @@ import contact from "../../assets/contact.png";
 import reject from "../../assets/reject.png"
 import heart from "../../assets/heart.png"
 import clean from "../../assets/clean.png"
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 function Home(props) {
 
@@ -15,13 +17,26 @@ function Home(props) {
         axios
             .get("https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gustavo/person")
             .then((res) =>{
-                console.log(res.data.profile)
                 setProfiles(res.data.profile)
+                // toastEmmitter()
             })
             .catch((err) =>{
                 console.log(err.message)
             })
     },[matches])
+
+    const toastEmmitter = () => {
+
+        toast('Woww deu Match!!🔥🔥', {
+            position: "top-center",
+            autoClose: 400,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            });
+    }
 
 
     const choosePerson = (idProfile, choiceProfile) =>{
@@ -34,37 +49,36 @@ function Home(props) {
         .post("https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gustavo/choose-person", body)
         .then((res)=>{
             setMatches(res.data)
+            toastEmmitter()
         })
         .catch((err)=>{
             console.log(err.message)
         })
     }
 
-    const clear = () =>{
-        axios
-        .put("https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gustavo/clear")
-        .then((res)=>{
-            console.log(res)
-        })
-        .catch((err)=>{
-            console.log(err.response)
-        })
-    }
+    // const clear = () =>{
+    //     axios
+    //     .put("https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gustavo/clear")
+    //     .then((res)=>{
+    //         alert('deletado com sucesso')
+    //     })
+    //     .catch((err)=>{
+    //         console.log(err.response)
+    //     })
+    // }
 
 
   return (
     <C.AppContainer>
     <C.MenuBar>
-        <button onClick={()=>clear()}><img src={clean} height ="35" width="35"></img></button>
+        <button><img src={clean} height ="35" width="35"></img></button>
         <h2>Astromatch</h2>
         <button onClick={props.goToMatches}><img src={contact} height ="45" width="45"></img></button>
     </C.MenuBar>
     {profiles === null ? (
-        
         <h1>Nada foi encontrado!</h1>
     ):(
         <C.Elements>
-        
         <C.Picture>
             <img src={profiles.photo}/>
         </C.Picture>
@@ -72,6 +86,7 @@ function Home(props) {
         <h2>🟢{profiles.name}, {profiles.age}</h2>
         <p>{profiles.bio}</p>
     </C.Profile>
+    <ToastContainer/>
         <C.Buttons>
             <button onClick={()=>choosePerson(profiles.id, false)}> <img src={reject} height ="30" width="30"></img></button>
             <button onClick={()=>choosePerson(profiles.id, true)}> <img src={heart} height ="30" width="30"></img></button>
