+++ modulo3/astromatch/src/components/Matches/Matches.js
@@ -3,6 +3,9 @@ import axios from 'axios'
 import * as C from './MatchesStyles';
 import clean from '../../assets/clean.png'
 import back from '../../assets/back.png'
+import Swal from "sweetalert2";
+// import { ToastContainer, toast } from 'react-toastify';
+// import 'react-toastify/dist/ReactToastify.css';
 
 function Matches(props) {
 
@@ -12,13 +15,26 @@ function Matches(props) {
     axios
     .get("https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gustavo/matches")
     .then((res)=>{
-        console.log(res.data.matches)
         setList(res.data.matches)
     })
     .catch((err)=>{
       console.log(err)
     })
   },[])
+
+//   const toastEmmitter = () => {
+
+//     toast('Woww deu Match!!🔥🔥', {
+//         position: "top-center",
+//         autoClose: 400,
+//         hideProgressBar: false,
+//         closeOnClick: true,
+//         pauseOnHover: true,
+//         draggable: true,
+//         progress: undefined,
+//         });
+// }
+
 
   const maping = list.map((profile)=>{
     return (
@@ -31,16 +47,34 @@ function Matches(props) {
     )
   })
 
-  return (
-    <C.ListContainer>
-        <C.ListMenuBar>
-            <button><img src={clean} height ="35" width="35"></img></button>
-            <h2>Astromatch</h2>
-            <button onClick={props.goToHome}><img src={back} height ="35" width="35"></img></button>
-        </C.ListMenuBar>
-        {maping}
-    </C.ListContainer>
-  )
-}
+  const clear = () =>{
+
+    axios
+    .put("https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gustavo/clear")
+    .then((res)=>{
+        window.location.reload(false);
+    })
+    .catch((err)=>{
+        console.log(err.response)
+      })
+
+      Swal.fire(
+        'Matches deletados!',
+        '',
+        'success')
+    }
+
+    return (
+      <C.ListContainer>
+        {/* <ToastContainer/> */}
+          <C.ListMenuBar>
+              <button onClick={clear}><img src={clean} height ="35" width="35"></img></button>
+              <h2>Astromatch</h2>
+              <button onClick={props.goToHome}><img src={back} height ="35" width="35"></img></button>
+          </C.ListMenuBar>
+          {maping}
+      </C.ListContainer>
+    )
+  }
 
 export default Matches;
